added test files

+++ Project_v2/app/src/main/java/com/example/projectv2/Model/User.java
@@ -106,7 +106,7 @@ public class User {
         this.organizerNotif = organizerNotif;
     }
 
-    public void setPhoneNumber(int phoneNumber) {
+    public void setPhoneNumber(long phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
     public void setProfileImage(String imageLink){
